Finish module orders and distribuitors and report module with charts, refactoring to architecture all app, pending using role
