+++ resources/views/distribuitorsTemplates/distribuitorPanel.blade.php
@@ -1,0 +1,5 @@
+<div id="map" style="width: 100%; height: 500px;">
+    <iframe
+        src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d63587.44873425483!2d-75.5186627813441!3d5.068782055284819!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x8e476ffa6a42ce3b%3A0xa863cf6423ea141c!2sManizales%2C%20Caldas!5e0!3m2!1ses!2sco!4v1643945844803!5m2!1ses!2sco"
+        width="100%" height="450" style="border:0;" allowfullscreen="" loading="lazy"></iframe>
+</div>
