+++ resources/views/helpers/messages.blade.php
@@ -1,0 +1,8 @@
+<div id="messageSuccess" class="alert alert-success d-flex align-items-center d-none" role="alert">
+    <svg class="bi flex-shrink-0 me-2" width="24" height="24" role="img" aria-label="Success:">
+        <use xlink:href="#check-circle-fill" />
+    </svg>
+    <div id="messageSuccessText">
+        <!--by controller-->
+    </div>
+</div>
