+++ resources/views/inventoriesTemplates/modals/modalObservations.blade.php
@@ -1,0 +1,29 @@
+<!-- Modal -->
+<div class="modal fade" id="modalObservations" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h5 class="modal-title" id="exampleModalLabel">Historial</h5>
+                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+            </div>
+            <div class="modal-body">
+                <div id="modalObservationsContent" class="row mb-3 ">
+                    <div class="input-group justify-content-center">
+                        <table class="table">
+                            <caption>Tabla Historial</caption>
+                            <thead class="table-light">
+                                <tr>
+                                    <th scope="col">#</th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                <!--to-do from javaScript controller-->
+                            </tbody>
+                        </table>
+                    </div>
+                </div>
+
+            </div>
+        </div>
+    </div>
+</div>
