+++ resources/views/stocksTemplates/stockPanel.blade.php
@@ -1,0 +1,19 @@
+<table class="table" id="stockTable">
+    <caption>Disponibilidad</caption>
+    <thead class="table-light">
+        <tr>
+            <th scope="col">#</th>
+            <th scope="col">Referencia</th>
+            <th scope="col">Nombre</th>
+            <th scope="col">Unidades Disponibles</th>
+            <th scope="col">Ultima Actualización</th>
+            <th scope="col">Precio Por Unidad</th>
+            <th scope="col">Peso Por Unidad</th>
+        </tr>
+    </thead>
+    <tbody>
+        <!--to-do from javaScript controller-->
+    </tbody>
+</table>
+
+@include('stocksTemplates.modals.modalEditStock')
